Bug 1669149 p6: Change print settings stored as floats to be stored as doubles. r=jwatt
This more precisely matches what we get from the printer and use in layout.

Depends on D99808

Differential Revision: https://phabricator.services.mozilla.com/D99810

+++ widget/nsPrintSettingsService.cpp
@@ -988,8 +988,7 @@ nsresult nsPrintSettingsService::WritePrefDouble(const char* aPrefId,
   NS_ENSURE_ARG_POINTER(aPrefId);
 
   nsAutoCString str;
-  // We cast to a float so we only get up to 6 digits precision in the prefs.
-  str.AppendFloat((float)aVal);
+  str.AppendFloat(aVal);
   return Preferences::SetCString(aPrefId, str);
 }
